ex03-algorithm-selection/requirements.txt
pytest
pycodestyle
numpy
scikit-learn>=0.18.0,<0.22
pandas
liac-arff

# add your requirements below this line
